% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_g_model.R
\name{density_g_model}
\alias{density_g_model}
\title{The l_1 to l_4 derivative from the g-modeling method}
\usage{
density_g_model(x, k, pi_0, lambda_set, g_prior)
}
\arguments{
\item{x}{a sequence of chi-squared test statistics}

\item{k}{degrees of freedom}

\item{pi_0}{the proportion of the null}

\item{lambda_set}{the set of noncentrality values}

\item{g_prior}{the prior probability for the noncentrality values}
}
\value{
a list: the margianl density, and its first-to-fourth derivatives
}
\description{
The l_1 to l_4 derivative from the g-modeling method
}
