% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_LS.R
\name{density_LS}
\alias{density_LS}
\title{log-density gradient-parametric approach}
\usage{
density_LS(x)
}
\arguments{
\item{x}{a sequence of chi-squared test statistics}
}
\value{
a list: the first-to-fourth derivatives
}
\description{
log-density gradient-parametric approach
}
