% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_PLS.R
\name{density_PLS}
\alias{density_PLS}
\title{Penalized least-squares method}
\usage{
density_PLS(x, qq)
}
\arguments{
\item{x}{a squence of chi-squared test statistics}

\item{qq}{the quantile used for splines}
}
\value{
a list: the first and second density derivatives
}
\description{
Penalized least-squares method
}
